+++ CJBCheatsMenu/release-notes.md
@@ -20,6 +20,7 @@
 * Improved compatibility with Android (thanks to kdau!).
 * Simplified move speed option.
 * 'Grow crops' and 'grow trees' can now be bound to the same button to grow both at once.
+* Fixed open cheats menu not updated correctly when game window is resized.
 * Fixed invalid time when changing time from after midnight.
 * Fixed instant-grown wild seeds being different from slow-grown ones (e.g. not applying botanist bonus, not randomized, etc).
 * Fixed instant-grown crops not generating giant crops.